TIL Rails: Actioncable integrated websockets for rails
Action Cable - Integrated websockets for Rails

Today I found an interesting gem that we might be able to implement for
future notification and stuff

The gem is call actioncable, in short it is basically a gem that
integrates websockets with Rails.

It built on top of `faye-websocket` and `celluloid`.

The only downside is that the only Database that it support right now is
Redis only.

for more information about the usage visit https://github.com/rails/actioncable